internal/v1: restructure unit tests
Move all compose related tests to separate file.
